requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
certifi==2026.4.22
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.3
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
distro==1.9.0
Django==5.2.13
exceptiongroup==1.3.1
groq==1.2.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
msgpack==1.1.2
packaging==26.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyOpenSSL==26.0.0
python-decouple==3.8
redis==7.4.0
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.5
tomli==2.4.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
whitenoise==6.12.0
zope.interface==8.3
